Use i8 for add/subtract amounts

+++ src/bf.zig
@@ -1,5 +1,5 @@
 pub const Op = union(enum) {
-    add: i32,
+    add: i8,
     move: i32,
     jump_if_zero: u32,
     jump_back_if_non_zero: u32,
